% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{BendingMoment}
\alias{BendingMoment}
\title{Bending moment of 9 types of truck}
\usage{
BendingMoment(Trucktype, BridgeL, v, Hz, NoiseLevel)
}
\arguments{
\item{Trucktype}{9 types of trucks, with 1-9}

\item{BridgeL}{Length of bridge (m)}

\item{v}{velocity of truck (m/s)}

\item{Hz}{Observation frequency.}

\item{NoiseLevel}{Usually 5\%, 10\%, 15\%, 20\% noise to simulate the real data.}
}
\value{
A sequence of bending moments generated at mid-span of bridge, within BridgeL/v time, every 1/Hz seconds.
\item{IL}{Static Influence Line of Bridge}
}
\description{
When a truck travels on a bridges, we use static influence line to
calculate unit load bending moment. We assume truck velocity is v(m/s),
and Observation frequency is Hz,so We will get a sequence of bending moments
within BridgeL/v time, every 1/Hz seconds, at the mid-span of bridge.
}
\details{
We want to return a sequence of bending moments at BridgeL/v time, every 1/Hz seconds, at the mid-span of bridge.
and in order to make data real, we add Noiselevel into data, and we have 9 types of default trucks,
each has different alxes load and alxes spacings, which was indicated in internal data.
}
\examples{
BendingMoment(1,100,50,50,0.15)
}
\author{
Qian Shi
}
