% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL.R
\name{IL}
\alias{IL}
\title{Influence line equation}
\usage{
IL(TruckD, BridgeL, SpaceL)
}
\arguments{
\item{TruckD}{Truck's position at the bridge.(TruckD <= BridgeL)}

\item{BridgeL}{length of Bridge}

\item{SpaceL}{Spacings between axle loads}
}
\value{
Bending moment influence line by unit load
}
\description{
Static Influence Line of Bridge
}
\details{
The bending moment of the truck passing this bridge can be calculated by the static influence line.
The influence line equation for the bending moment of unit load can be expressed as IL function.
}
\author{
Qian Shi
}
